Convert demo services into TypeScript
