+++ frontend/package.json
@@ -36,7 +36,6 @@
   "devDependencies": {
     "@babel/core": "^7.20.0",
     "@babel/eslint-parser": "^7.25.1",
-    "@babel/preset-env": "^7.26.9",
     "@babel/preset-typescript": "^7.26.0",
     "@testing-library/jest-native": "^5.4.3",
     "@testing-library/react-native": "^12.9.0",
@@ -44,8 +43,6 @@
     "@types/react": "^18.3.18",
     "@types/react-native": "^0.72.8",
     "@types/react-router-native": "^5.1.3",
-    "@typescript-eslint/eslint-plugin": "^8.26.1",
-    "@typescript-eslint/parser": "^8.26.1",
     "babel-jest": "^29.7.0",
     "cross-env": "^7.0.3",
     "eslint": "^8.57.0",
@@ -54,7 +51,6 @@
     "eslint-plugin-react-native": "^4.1.0",
     "jest": "^29.7.0",
     "jest-expo": "~52.0.4",
-    "react-test-renderer": "^18.3.1",
     "ts-jest": "^29.2.6",
     "ts-node": "^10.9.2",
     "typescript": "^5.8.2"
