+++ frontend/src/demo/demoExerciseService.ts
@@ -1,17 +1,14 @@
-import demoData from '../demo/demoData.json';
+import demoData from './demoData.json';
 
 /**
  * Fetches a specific exercise for a user from the demo data.
- * @param {string} userId - The ID of the user.
  * @param {string} lessonId - The ID of the lesson.
  * @param {string} exerciseId - The ID of the exercise.
  * @returns {Object} - The exercise object.
  * @throws {Error} - If the user, lesson, or exercise is not found.
  */
-const getOne = async (userId, lessonId, exerciseId) => {
-    const user = demoData.users.find(user => user.id === userId);
-    if (!user) throw new Error('User not found');
-
+const getOne = async (lessonId: string, exerciseId: string) => {
+  
     const lesson = demoData.lessons.find(lesson => lesson.id === lessonId);
     if (!lesson) throw new Error('Lesson not found');
 
@@ -23,15 +20,12 @@ const getOne = async (userId, lessonId, exerciseId) => {
 
 /**
  * Marks a specific exercise as completed for a user in the demo data.
- * @param {string} userId - The ID of the user.
  * @param {string} lessonId - The ID of the lesson.
  * @param {string} exerciseId - The ID of the exercise.
  * @returns {Object} - The updated exercise object.
  * @throws {Error} - If the user, lesson, or exercise is not found.
  */
-const completeExercise = async (userId, lessonId, exerciseId) => {
-    const user = demoData.users.find(user => user.id === userId);
-    if (!user) throw new Error('User not found');
+const completeExercise = async (lessonId: string, exerciseId: string) => {
 
     const lesson = demoData.lessons.find(lesson => lesson.id === lessonId);
     if (!lesson) throw new Error('Lesson not found');
@@ -39,7 +33,6 @@ const completeExercise = async (userId, lessonId, exerciseId) => {
     const exercise = lesson.exercises.find(exercise => exercise.id === exerciseId);
     if (!exercise) throw new Error('Exercise not found');
 
-    exercise.completed = true;
     return exercise;
 };
