+++ frontend/src/demo/demoLoginService.ts
@@ -1,4 +1,5 @@
-import demoData from '../demo/demoData.json';
+import { LoginCredentials } from '../types';
+import demoData from './demoData.json';
 
 /**
  * Authenticates a user with the provided credentials.
@@ -8,7 +9,7 @@ import demoData from '../demo/demoData.json';
  * @returns {Object} - An object containing a token and the user data.
  * @throws {Error} - If the username or password is invalid.
  */
-const login = async (credentials) => {
+const login = async (credentials: LoginCredentials) => {
     const user = demoData.users.find(user => user.username === credentials.username && user.password === credentials.password);
     if (!user) throw new Error('Invalid username or password');
     return { token: 'demo-token', user };
